Create a set of Single instances
Associate a Single instance with user input and use "doOnError" method in
order to handle the errors caused by invalid values inserted in those
fields. Then, zip those instances in order to create a new Single instance
that performs all the work.
In this way, one can get rid of mutable variables in the Model and be able
to highlight a field that causes the error.